requirements.txt
click==6.6
Flask==0.11.1
flask-heroku==0.1.9
Flask-SQLAlchemy==2.1
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.2
slackclient==1.0.2
SQLAlchemy==1.1.2
synapse-pay-rest-native==1.0.2
Werkzeug==0.11.11
